requirements.txt
click==6.2
python-magic==0.4.10
